Make uacpi_kernel_get_ticks work with nanoseconds and rename
This used to return 100 nanosecond ticks, which is an AML implementation
detail and is of little importance for the client of the library. Handle
this internally instead and let the client return a regular nanosecond
counter for simplicity.

This also renames the API to uacpi_kernel_get_nanoseconds_since_boot.

Signed-off-by: Daniil Tatianin <99danilt@gmail.com>